+++ lib/services/migration/migration.service.d.ts
@@ -11,8 +11,7 @@ export declare class MigrationService {
     private logger;
     constructor(configService: ConfigService, database: DatabaseService, migrationsResolver: MigrationsResolver, logger: LogFactory);
     connect(): Promise<import("mongodb").Db>;
-    ensureFolders(): Promise<void>;
-    up(): Promise<any[]>;
+    up(): Promise<ReturnType[]>;
     down(): Promise<unknown>;
     createWithTemplate(template: TemplateTypes, name: string): Promise<string>;
     init(): Promise<void>;