+++ lib/templates/basic.d.ts
@@ -1,2 +1,0 @@
-declare const _default: "\nmodule.exports = {\n  async up () {\n    return ['Up']\n  },\n\n  async down () {\n    return ['Down']\n  }\n}\n";
-export default _default;