+++ lib/templates/es5.d.ts
@@ -1,0 +1,2 @@
+declare const _default: "\nmodule.exports = {\n  async up (db) {\n    return ['Up']\n  },\n\n  async down (db) {\n    return ['Down']\n  }\n}\n";
+export default _default;