+++ lib/templates/es6.d.ts
@@ -1,0 +1,2 @@
+declare const _default: "\nexport async function up(db) {\n  return ['Up'];\n}\nexport async function down(db) {\n  return ['Down'];\n}\n";
+export default _default;