+++ lib/templates/index.d.ts
@@ -1,6 +1,7 @@
 import native from './native';
-import basic from './basic';
+import es5 from './es5';
+import es6 from './es6';
 import typescript from './typescript';
 import migration from './migration';
-export { basic, native, typescript, migration };
-export declare type TemplateTypes = 'basic' | 'native' | 'typescript' | 'migration';
+export { es6, es5, native, typescript, migration };
+export declare type TemplateTypes = 'es5' | 'es6' | 'native' | 'typescript' | 'migration';