+++ src/templates/es6.ts
@@ -1,0 +1,8 @@
+export default `
+export async function up(db) {
+  return ['Up'];
+}
+export async function down(db) {
+  return ['Down'];
+}
+`;