+++ src/templates/index.ts
@@ -1,7 +1,8 @@
 import native from './native';
-import basic from './basic';
+import es5 from './es5';
+import es6 from './es6';
 import typescript from './typescript';
 import migration from './migration';
 
-export { basic, native, typescript, migration };
-export type TemplateTypes = 'basic' | 'native' | 'typescript' | 'migration';
+export { es6, es5, native, typescript, migration };
+export type TemplateTypes = 'es5' | 'es6' | 'native' | 'typescript' | 'migration';