+++ src/templates/native.ts
@@ -1,21 +1,19 @@
 export default `
-module.exports = {
-  async up (db) {
-    await db
-      .collection('albums')
-      .updateOne({ artist: 'The Beatles' }, { $set: { blacklisted: true } })
-    await db
-      .collection('albums')
-      .updateOne({ artist: 'The Doors' }, { $set: { stars: 5 } })
-  },
+export async function up (db) {
+  await db
+    .collection('albums')
+    .updateOne({ artist: 'The Beatles' }, { $set: { blacklisted: true } })
+  await db
+    .collection('albums')
+    .updateOne({ artist: 'The Doors' }, { $set: { stars: 5 } })
+},
 
-  async down (db) {
-    await db
-      .collection('albums')
-      .updateOne({ artist: 'The Doors' }, { $set: { stars: 0 } })
-    await db
-      .collection('albums')
-      .updateOne({ artist: 'The Beatles' }, { $set: { blacklisted: false } })
-  }
+export async function down (db) {
+  await db
+    .collection('albums')
+    .updateOne({ artist: 'The Doors' }, { $set: { stars: 0 } })
+  await db
+    .collection('albums')
+    .updateOne({ artist: 'The Beatles' }, { $set: { blacklisted: false } })
 }
 `;